+++ fabric/src/main/java/com/railwayteam/railways/fabric/mixin/FluidHelperMixin.java
@@ -1,0 +1,37 @@
+/*
+ * Steam 'n' Rails
+ * Copyright (c) 2026 The Railways Team
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.railwayteam.railways.fabric.mixin;
+
+import com.google.gson.JsonObject;
+import com.railwayteam.railways.util.FluidUtils;
+import com.simibubi.create.foundation.fluid.FluidHelper;
+import io.github.fabricators_of_create.porting_lib.fluids.FluidStack;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+// mods like KubeJS bypass the ProcessingRecipeSerializer implementation, so this is a backup demangler
+@Mixin(FluidHelper.class)
+public class FluidHelperMixin {
+    @Inject(method = "deserializeFluidStack", at = @At("HEAD"), remap = false)
+    private static void demangle(JsonObject json, CallbackInfoReturnable<FluidStack> cir) {
+        FluidUtils.demangleFluidAmount(json);
+    }
+}